requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2
python-decouple==3.8
django-cors-headers==4.3.1
Pillow
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
boto3==1.34.0
django-filter==23.5
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0